requirements.txt
asgiref==3.6.0
certifi==2022.12.7
charset-normalizer==3.1.0
cvzone==1.5.6
Django==4.1.6
django-countries==7.5.1
idna==3.4
numpy==1.24.2
opencv-python==4.7.0.72
Pillow==9.4.0
pyotp==2.8.0
requests==2.28.2
six==1.16.0
sqlparse==0.4.3
typing_extensions==4.5.0
tzdata==2022.7
urllib3==1.26.14
